Migrate most changes from Grammar/Grammar to lib2to3's copy.
This started with a wholesale copy of the current version, after which I
manually ported the edits required for:
- print statements
- exec statements
- lambdas without parens
- tuple unpacking in arguments
- commas in raise/except
- backtick expressions

Many of the symbols changed names/hierarchy, and this is a breaking change for
many users of lib2to3.  Some existing patterns will not even compile, but more
frustrating, some will but they will not work (e.g. `comp_for` or `power`)

There is still one failing test related to has_key, that I have not figured out.